Factor-out logging, run_gdal_subcommand stuff in batch scripts
